requirements.txt
scanpy
magic-impute
numpy
pandas
matplotlib
scikit-learn
deepimpute
scvi-tools